[fUML4CPP] initialization of execution plugin is moved to a special operation named "initialized"
Because in MDE4CPP ALL plugins (Dlls) existing in the application/bin directory are started when executing an arbitrary model, the initialization of every plugin should be avoided.
For that, initialization of an execution plugin is now handled in a new operation called "initialize()" which is called inside "executeActivity()". The flag "m_IsInitialized" ensures that initialization of the plugin is only done once.